// src/buildpc/interfaces/buildpc.ts
export interface BuildPC {
    mothers?:     Mother[];
    cpu?:         CPU[];
    coolercpu?:   Coolercpu[];
    mem?:         Mem[];
    video?:       Video[];
    hd?:          HD[];
    gab?:         Gab[];
    fuente?:      Fuente[];
    monitores?:   Monitore[];
    auriculares?: Auriculare[];
    card_reader?: any[];
    coolerfan?:   Coolerfan[];
    mouse?:       Mouse[];
    mousepad?:    Mousepad[];
    opticos?:     any[];
    os?:          any[];
    soportes?:    any[];
    teclados?:    Teclado[];
    extras?:      any[];
    notebooks?:   Notebook[];
}

export interface Auriculare {
    id_producto:         number;
    disponible:          number;
    mini_plug_audio:     null | string;
    mini_plug_combinado: null | string;
    impedancia:          null | string;
    usb:                 string;
    wifi_por_receptor:   string;
    wifi_por_bt:         string;
    tamano_parlante:     string;
    tipo:                string;
    resp_hz_min:         string;
    resp_hz_max:         string;
    sensibilidad:        null | string;
    colores:             Colores;
    mini_plug_mic:       null | string;
    usb_power:           string;
    disp_compatible:     string;
    largo_cable:         null | string;
    iluminacion:         ColorLuzTecladoEnum | null;
    base:                string;
    mic_resp_hz_min:     null | string;
    mic_resp_hz_max:     null | string;
    mic_sensibilidad:    null;
    mic_tipo:            MicTipo | null;
    mic_cancelacion:     string;
    mic_tipo_captacion:  MicTipoCaptacion | null;
    surround:            string;
    wireless:            Wireless;
    con_mic:             string;
    virtual_real:        VirtualReal;
    formato:             Formato;
}

export enum Colores {
    Blanco = "Blanco",
    Negro = "Negro",
    NegroAmarillo = "Negro,Amarillo",
    NegroRojo = "Negro,Rojo",
    Rosa = "Rosa",
}

export enum Formato {
    Headset = "Headset",
}

export enum ColorLuzTecladoEnum {
    RGB = "RGB",
    Rojo = "Rojo",
    SinIluminación = "Sin Iluminación",
}

export enum MicTipo {
    Desmontable = "Desmontable",
    Integrado = "Integrado",
    Retráctil = "Retráctil",
}

export enum MicTipoCaptacion {
    NoEspecificado = "No especificado",
    Omnidireccional = "Omnidireccional",
    Unidireccional = "Unidireccional",
}

export enum VirtualReal {
    Real = "Real",
    Virtual = "Virtual",
}

export enum Wireless {
    Cableado = "Cableado",
    Inalámbrico = "Inalámbrico",
}

export interface Coolercpu {
    id_producto:       number;
    sockets:           string;
    disponible:        number | null;
    altura:            number;
    tipo_gab:          TipoGab;
    espacios_120mm:    string;
    espacios_140mm:    number;
    altura_real:       string;
    nro_coolers:       number;
    w:                 number;
    cooler_base:       number;
    tipo_disipacion:   TipoDisipacion | null;
    color_iluminacion: null | string;
    ruido:             null;
    ruido_max:         null | string;
    sistema_rgb:       CoolercpuSistemaRGB | null;
    TDP:               null | string;
    tam_cooler:        null | string;
}

export enum CoolercpuSistemaRGB {
    ARGBHeader = "ARGB Header",
    Empty = "",
    NoIncluye = "No incluye",
    PropioDelGabinete = "Propio del gabinete",
    RGBHeader = "RGB Header",
}

export enum TipoDisipacion {
    Aire = "Aire",
    Watercooling = "Watercooling",
}

export enum TipoGab {
    Atx = "ATX",
    EAtx = "E-ATX",
    Itx = "ITX",
    MAtx = "M-ATX",
}

export interface Coolerfan {
    id_producto:              number;
    disponible:               number;
    dimensiones:              number;
    cantidad:                 number;
    max_speed:                number | null;
    regular_speed:            number;
    LED:                      number;
    ColorLED:                 string;
    molex:                    number;
    Pines:                    number;
    max_ruido:                number | null;
    flujo_maximo:             null;
    espesor:                  number;
    hue_compatible:           number;
    w:                        number;
    presion_estatica:         number | null;
    sistema_rgb:              CoolercpuSistemaRGB | null;
    incluye_controladora_rgb: number;
}

export interface CPU {
    id_producto:         number;
    modelo:              string;
    cores:               number;
    disponible:          number;
    w:                   number;
    clock:               number;
    tipo:                string;
    cooler:              number;
    familia:             string;
    l1:                  null | string;
    l2:                  null | string;
    l3:                  null | string;
    tecnologia:          string;
    cate_cpu:            number;
    chipset_gpu:         string;
    nivel_cpu_deprecado: null;
    nivel_gpu:           number;
    hilos:               number;
    turbo:               number | null;
}

export interface Fuente {
    id_producto:      number;
    disponible:       number;
    w_real:           number;
    w_nominal:        number;
    tipo:             TipoFuenteEnum;
    marca:            null;
    _24pines:         number;
    _4pines:          number;
    _4plus:           number;
    pines_pcie:       number;
    pines_pcieplus:   number;
    incluye_cable:    number;
    kit:              number;
    compatible_abajo: number;
    certi80plus:      TresD | null;
    ModoHibrido:      number;
    tipocableado:     Tipocableado;
    amper12v:         number | null;
    sata:             number;
    molex:            string;
    floppy:           null;
    digital:          number;
    color:            Color;
    iluminacion:      ColorLuzTecladoEnum | null;
}
// ...
export interface Mem {
    id_producto:     number;
    capacidad:       number;
    disponible:      number;
    disipador:       number;
    velocidad:       string;
    tipo:            PurpleTipo;
    w:               number;
    slot:            number;
    sodimm:          number;
    activo:          number;
    latencia_cl:     number;
    voltaje:         number;
    marca:           string;
    disipacion_alta: number;
}
// ...
export interface Mother {
    id_producto:     number;
    modelo:          null | string;
    cpus:            string;
    SATA:            number;
    marca:           null | string;
    _4pines:         number;
    _4plus:          number;
    wmax:            number;
    _24pines:        number;
    PCI:             number;
    nro_pcie:        number;
    PCIE1X:          number;
    nro_agp:         number;
    IDE:             number;
    tipo_multiplaca: string;
    DDR:             number;
    DDR2:            number;
    DDR3:            number;
    DDR4:            number;
    DDR5:            number;
    Chipsets:        string;
    disponible:      number;
    vga:             number;
    HDMI:            number;
    dvi:             string;
    DP:              number;
    w:               number;
    Floppy:          number;
    USB:             number;
    Factor:          TipoGab;
    sodimm:          number;
    m2:              number;
    con_micro:       number;
    id_sub_cate:     number;
    tipo:            TipoCPUEnum;
    activo:          number;
    wifi_integrada:  number;
    sistema_rgb:     MotherSistemaRGB | null;
    usb20_trasero:   number;
    usb30_trasero:   number;
    usb31_trasero:   number;
    ps2:             number;
    lan:             LAN;
    placa_audio:     string;
    anclaje:         string;
    bios_flashback:  number;
    usb_32_trasero:  number;
    usb_type_c:      number;
    pci_e_x4:        number;
    m2_sata:         number;
    m2_nvme:         number;
    tipo_memoria:    PurpleTipo;
    slot_memoria:    number;
}
// ...
export interface Mouse {
    id_producto:             number;
    disponible:              number;
    color:                   string;
    cantidad_de_botones:     number;
    dpi_ajustable:           string;
    dpi_minimo:              null | string;
    dpi_maximo:              number;
    frecuencia_de_respuesta: null | string;
    peso:                    null | string;
    sistema_de_pesas:        string;
    peso_de_las_pesas:       null | string;
    cantidad_de_pesas:       null | string;
    ancho:                   null | string;
    profundidad:             null | string;
    alto:                    null | string;
    tipo_switch:             TipoSwitch | null;
    tipo_sensor:             TipoSensor | null;
    sensor:                  null | string;
    tecnologias_compatibles: null;
    iluminacion:             string;
    bateria_interna:         string;
    pilas:                   null | string;
    inalambrico:             number;
    tipo_cable:              MouseTipoCable | null;
    cable_largo:             null | string;
    cable_desmontable:       string;
    teconologiawifi:         null | string;
    usb_requeridos:          null | string;
    incluye_receptor:        string;
    orientacion:             Orientacion | null;
    ps2_requerido:           string;
    sistema_recargable:      string;
    velocidad_maxima:        number | null;
    aceleracion_maxima:      number | null;
    funda:                   string;
    consolas_compatibles:    null;
    base_de_carga:           string;
    iluminancion_colores:    IluminancionColores | null;
    procesador:              null | string;
    tipo_de_agarre:          null | string;
    agarre_palm_grip:        number;
    agarre_claw_grip:        number;
    agarre_finger_grip:      number | null;
    agarre_ergonomico:       number | null;
    en_combo_con_pad:        number | null;
}

export enum IluminancionColores {
    Azul = "Azul",
    RGB = "RGB",
    Rainbow = "Rainbow",
    SinIluminación = "Sin Iluminación",
}

export enum Orientacion {
    ParaAmbidiestros = "Para ambidiestros",
    ParaDiestros = "Para diestros",
}

export enum MouseTipoCable {
    ParacordUltraLiviano = "Paracord Ultra Liviano",
    Plástico = "Plástico",
    TelaMallada = "Tela mallada",
}

export enum TipoSensor {
    Óptico = "Óptico",
}

export enum TipoSwitch {
    Omron = "Omron",
}

export interface Mousepad {
    id_producto:     number;
    disponible:      number;
    ancho:           number;
    alto:            number;
    grosor:          number;
    tipo_superficie: TipoSuperficie;
    tipo:            MousepadTipo;
    iluminacion:     ColorLuzTecladoEnum | null;
    categoria:       string;
    cargas_wifi:     string;
    hub_usb:         string;
    req_usb:         string;
    control_Led:     MicTipo | null;
}
// ...
export interface Teclado {
    id_producto:                   number;
    disponible:                    number;
    swich_type:                    null | string;
    switch_spe:                    null | string;
    tipo_teclado:                  TipoTeclado;
    iluminacion:                   TecladoIluminacion | null;
    con_pad_numerico:              string;
    tecnologias_compatibles:       TecnologiasCompatibles | null;
    puertos_usb_hb:                null | string;
    tipo_cable:                    TecladoTipoCable | null;
    cable_extraible:               string;
    ancho:                         string;
    profundidad:                   string;
    alto:                          string;
    antisalpicadura:               string;
    idioma:                        Idioma | null;
    distribucion:                  Distribucion | null;
    antighosting:                  string;
    peso:                          null | string;
    reposamunecas:                 string;
    reposamunecas_extraible:       string;
    color:                         string;
    cantidad_macros_cargar:        null | string;
    cantidad_teclas_en_simultaneo: null | string;
    conexion:                      string;
    puertos_ps2:                   number;
    bluetooth:                     number;
    receptor_bluetooth:            number;
    wifi:                          string;
    bateria:                       Bateria | null;
    tipo_mecanicos:                TipoMecanicos;
    NKey_Rollover:                 string;
    audio_hub:                     string;
    funda:                         number;
    material:                      Material;
    polling_rate:                  null | string;
    procesador:                    null | string;
    tiempo_de_respuesta:           null | string;
    memoria:                       null | string;
    largo_del_cable:               null | string;
    touchpad:                      string;
    alcance_wifi:                  null;
    pilas_incluida:                null | string;
    ancho_sin_pad:                 null;
}

export enum Bateria {
    BateríaInterna = "Batería interna",
    Empty = "",
    The2PilasAAA = "2 pilas AAA",
}

export enum Distribucion {
    España = "España",
    EstadosUnidos = "Estados Unidos",
    Latinoamerica = "Latinoamerica",
}

export enum Idioma {
    Español = "Español",
    Inglés = "Inglés",
}